Refactor hyper|history tree stats and use expvars
Create new metrics package and use expvars to recollect stats.
Remove old stats files.
