--- balloon/hyper/stats.go
@@ -1,30 +1,0 @@
-/*
-   Copyright 2018 Banco Bilbao Vizcaya Argentaria, S.A.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-
-package hyper
-
-// stats to measure the performance of the tree and
-// the associated caches and strategies
-type stats struct {
-	hits   int
-	disk   int
-	dh     int
-	update int
-	leaf   int
-	lh     int
-	ih     int
-	lend   float64
-}
